feat: Swap Context
