+++ src/components/swap/SwapSkeleton.tsx
@@ -1,10 +1,9 @@
 import { Trans } from '@lingui/macro'
-import { ArrowContainer } from 'pages/Swap'
 import { ArrowDown } from 'react-feather'
 import styled, { useTheme } from 'styled-components'
 import { ThemedText } from 'theme/components'
 
-import { ArrowWrapper } from './styled'
+import { ArrowContainer, ArrowWrapper } from './styled'
 
 const StyledArrowWrapper = styled(ArrowWrapper)`
   position: absolute;
