+++ src/components/swap/styled.tsx
@@ -154,3 +154,52 @@ export const SwapShowAcceptChanges = styled(AutoColumn)`
   border-radius: 12px;
   margin-top: 8px;
 `
+
+export const SwapSection = styled.div`
+  background-color: ${({ theme }) => theme.surface2};
+  border-radius: 16px;
+  color: ${({ theme }) => theme.neutral2};
+  font-size: 14px;
+  font-weight: 500;
+  height: 120px;
+  line-height: 20px;
+  padding: 16px;
+  position: relative;
+
+  &:before {
+    box-sizing: border-box;
+    background-size: 100%;
+    border-radius: inherit;
+
+    position: absolute;
+    top: 0;
+    left: 0;
+
+    width: 100%;
+    height: 100%;
+    pointer-events: none;
+    content: '';
+    border: 1px solid ${({ theme }) => theme.surface2};
+  }
+
+  &:hover:before {
+    border-color: ${({ theme }) => theme.deprecated_stateOverlayHover};
+  }
+
+  &:focus-within:before {
+    border-color: ${({ theme }) => theme.deprecated_stateOverlayPressed};
+  }
+`
+
+export const OutputSwapSection = styled(SwapSection)`
+  border-bottom: ${({ theme }) => `1px solid ${theme.surface1}`};
+`
+
+export const ArrowContainer = styled.div`
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 100%;
+  height: 100%;
+`
