+++ src/utils/percent.ts
@@ -1,0 +1,12 @@
+import { Percent } from '@uniswap/sdk-core'
+
+export function largerPercentValue(a?: Percent, b?: Percent) {
+  if (a && b) {
+    return a.greaterThan(b) ? a : b
+  } else if (a) {
+    return a
+  } else if (b) {
+    return b
+  }
+  return undefined
+}
